requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pytest>=7.4.0
tqdm>=4.65.0
pyyaml>=6.0.0
requests>=2.30.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
websocket-client>=1.6.0
sqlalchemy>=2.0.20
redis>=5.0.0
pymongo>=4.5.0

# Note: The following packages may need manual installation
ta-lib>=0.4.25  # Required: See ENVIRONMENT_SETUP.md for installation instructions
# alpaca-trade-api>=3.0.0  # Optional
# ccxt>=4.0.0  # Optional
mcp-server>=0.1.4  # Required for MCP Server Configuration
mcp-client==0.0.0  # Required for MCP Server Configuration

# Additional dependencies
pydantic>=2.4.0
sse-starlette>=1.6.0
aiohttp>=3.8.0
asyncio>=3.4.0
httpx>=0.24.0
python-multipart>=0.0.6
